There are more changes on this commit than there should be.   Adding a test for an appointment which had one hour of normal time, one hour of bedtime and one hour of overtime uncovered flaws in the code particulary around bedtime pay calc where the distinction between bt abd ot where not being handled correctly.
As a result, I developed a bit of panic mode and used println on variuables until the code was fixed and all the tests passed.  I hope this helps explain my thought process a little better for the times you see more code changes commited than should be.
I strive to achieve proper test driven from but it is a work in progress.

Fixed typos in the tests to correct all times to 24 hour format.

all tests passed.